Add `-p` (`--publicDirectory`) option to the `favicons` command
Favicons are typically stored in a dedicated directory for static assets rather than in `src/shared`. In most frameworks, this directory is `public` by default. Therefore, it’s better for the `favicons` command to use this directory by default. A separate option is needed to specify a directory other than `public` because this cannot be implemented within the `-i` and `-o` options. This allows for the execution of the general `all` command.